Support messages and repairs larger than 2GB:  Length field in PeerChannel header is now 8 bytes.  Incremented minor protocol version from 3 to 4.  Use FilePeerMessage for large outgoing messages.

+++ test/src/org/lockss/poller/v3/FuncV3Voter.java
@@ -378,7 +378,7 @@ public class FuncV3Voter extends LockssTestCase {
     public PeerMessage newPeerMessage() {
       throw new UnsupportedOperationException("Not implemented");
     }
-    public PeerMessage newPeerMessage(int estSize) {
+    public PeerMessage newPeerMessage(long estSize) {
       throw new UnsupportedOperationException("Not implemented");
     }
     public void registerMessageHandler(int protocol, MessageHandler handler) {
